Added BigInt lib

+++ include/Blocks/blocks.h
@@ -7,6 +7,7 @@
 
 #include <stdlib.h>
 #include <stdbool.h>
+#include "bigint.h"
 
 // BlockKeeper
 /*
@@ -24,7 +25,7 @@
 typedef struct blockNode {
     unsigned int lineStarted;
     long long int fPointer;
-    long long int *loopCounter;
+    BigInt *loopCounter;
     bool isIntConstant;
     bool isShortHandLoop;
     struct blockNode *nests;
@@ -36,7 +37,7 @@ typedef struct blockKeeper {
 } BlockKeeper;
 
 
-void openBlock(BlockKeeper *blockKeeper, long long int *loopCounter, unsigned int line, unsigned long long int fPointer, bool isIntConstant, bool isShortHandLoop);
+void openBlock(BlockKeeper *blockKeeper, BigInt *loopCounter, unsigned int line, unsigned long long int fPointer, bool isIntConstant, bool isShortHandLoop);
 
 Block* getBlock(BlockKeeper* blockKeeper);
